Fixed error caused by non-inclusion of cc2420_txrx() prototype

+++ boards/telosb/driver_cc2420.c
@@ -16,6 +16,7 @@
 #include "crash.h"
 
 #include "cc2420.h"
+#include "cc2420_arch.h"
 
 #define ENABLE_DEBUG    (1)
 #include "debug.h"
@@ -100,7 +101,7 @@ int cc2420_get_fifop(void)
     return CC2420_FIFOP;
 }
 
-int cc2420_get_sfd(void)
+uint8_t cc2420_get_sfd(void)
 {
     return CC2420_SFD;
 }
